+++ NEWS.md
@@ -1,3 +1,6 @@
+* Added option to exclude outbreak-associated cases from baseline fitting for GLM-based algorithms
+* Added visualisation of age group comparison of cases in signal vs. cases from the signal detection period in the linelist tab
+
 # SignalDetectionTool 0.10.0
 
 * Bug fixes and improvements to the HTML report
@@ -18,8 +21,6 @@
 * Added option to extend the linelist up to the current day
 * Added new multi-seasonal harmonic method with annual and semiannual components and optional timetrend
 * Added handling for ages and age-group bounds greater than or equal to 115 to avoid unplausible ages in linelist
-* Added visualisation of age group comparison of cases in signal vs. cases from the signal detection period in the linelist tab
-* integrated missing ISO calendar week handling (add_cw_iso()) directly into aggregate_data(), removing the need for a seperate preprocessing step
 
 # SignalDetectionTool 0.9.0